Styles Refactor
