+++ .all-contributorsrc
@@ -45,6 +45,15 @@
       "contributions": [
         "code"
       ]
+    },
+    {
+      "login": "Laxmankohar",
+      "name": "Laxman kohar",
+      "avatar_url": "https://avatars.githubusercontent.com/u/73365025?v=4",
+      "profile": "https://github.com/Laxmankohar",
+      "contributions": [
+        "code"
+      ]
     }
   ],
   "contributorsPerLine": 7,
