+++ admin/src/components/Attendence/ViewAttendence.tsx
@@ -1,14 +1,5 @@
 import { createStyles, Table, Progress, Anchor, Text, Group, ScrollArea } from '@mantine/core';
-
-const useStyles = createStyles((theme) => ({
-  progressBar: {
-    '&:not(:first-of-type)': {
-      borderLeft: `3px solid ${theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.white}`,
-    },
-  },
-}));
-
-
+import useStyles from './style';
 const data = [
       {
         "title": "Foundation",
