+++ admin/src/components/Attendence/style.ts
@@ -1,0 +1,13 @@
+import { createStyles } from "@mantine/core";
+
+const useStyles = createStyles((theme) => ({
+  progressBar: {
+    "&:not(:first-of-type)": {
+      borderLeft: `3px solid ${
+        theme.colorScheme === "dark" ? theme.colors.dark[7] : theme.white
+      }`,
+    },
+  },
+}));
+
+export default useStyles;
