+++ admin/src/components/Navbar/Navbar.tsx
@@ -11,6 +11,7 @@ import {
 } from "@tabler/icons";
 import { UserButton } from "../Navbar/UserButton";
 import { LinksGroup } from "../Navbar/NavbarLinksGroup";
+import useStyles from "./NavbarStyle";
 
 const mockdata = [
   { label: "Dashboard", icon: IconGauge, link: "/" },
@@ -48,86 +49,6 @@ const mockdata = [
     ],
   },
 ];
-
-const useStyles = createStyles((theme, _params, getRef) => {
-  const icon = getRef("icon");
-  return {
-    navbar: {
-      backgroundColor:
-        theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.white,
-      paddingBottom: 0,
-    },
-
-    header: {
-      padding: theme.spacing.md,
-      paddingTop: 0,
-      marginLeft: -theme.spacing.md,
-      marginRight: -theme.spacing.md,
-      color: theme.colorScheme === "dark" ? theme.white : theme.black,
-      borderBottom: `1px solid ${
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[4]
-          : theme.colors.gray[3]
-      }`,
-    },
-
-    links: {
-      marginLeft: -theme.spacing.md,
-      marginRight: -theme.spacing.md,
-    },
-
-    link: {
-      ...theme.fn.focusStyles(),
-      display: "flex",
-      alignItems: "center",
-      textDecoration: "none",
-      fontSize: theme.fontSizes.sm,
-      color:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[1]
-          : theme.colors.gray[7],
-      padding: `${theme.spacing.xs}px ${theme.spacing.sm}px`,
-      borderRadius: theme.radius.sm,
-      fontWeight: 500,
-
-      "&:hover": {
-        backgroundColor:
-          theme.colorScheme === "dark"
-            ? theme.colors.dark[6]
-            : theme.colors.gray[0],
-        color: theme.colorScheme === "dark" ? theme.white : theme.black,
-
-        [`& .${icon}`]: {
-          color: theme.colorScheme === "dark" ? theme.white : theme.black,
-        },
-      },
-    },
-
-    linkIcon: {
-      ref: icon,
-      color:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[2]
-          : theme.colors.gray[6],
-      marginRight: theme.spacing.sm,
-    },
-
-    linksInner: {
-      paddingTop: theme.spacing.xl,
-      paddingBottom: theme.spacing.xl,
-    },
-
-    footer: {
-      marginLeft: -theme.spacing.md,
-      marginRight: -theme.spacing.md,
-      borderTop: `1px solid ${
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[4]
-          : theme.colors.gray[3]
-      }`,
-    },
-  };
-});
 
 export function NavbarNested() {
   const { classes, cx } = useStyles();
