+++ client/src/components/MainHeader/MainHeader.tsx
@@ -4,50 +4,7 @@ import { useDisclosure } from '@mantine/hooks';
 import { MantineLogo } from '@mantine/ds';
 import DarkThemeToggle from '../DarkThemeToggle/DarkThemeToggle';
 import { NavLink } from 'react-router-dom';
-
-const useStyles = createStyles((theme) => ({
-  header: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    height: '100%',
-  },
-
-  links: {
-    [theme.fn.smallerThan('xs')]: {
-      display: 'none',
-    },
-  },
-
-  burger: {
-    [theme.fn.largerThan('xs')]: {
-      display: 'none',
-    },
-  },
-
-  link: {
-    // display: 'block',
-    // lineHeight: 1,
-    // padding: '8px 12px',
-    // borderRadius: theme.radius.sm,
-    // textDecoration: 'none',
-    color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[7],
-    fontSize: theme.fontSizes.sm,
-    fontWeight: 500,
-
-    '&:hover': {
-      cursor: 'pointer',
-      backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
-    },
-  },
-
-  linkActive: {
-    '&, &:hover': {
-      backgroundColor: theme.fn.variant({ variant: 'light', color: theme.primaryColor }).background,
-      color: theme.fn.variant({ variant: 'light', color: theme.primaryColor }).color,
-    },
-  },
-}));
+import useStyles from './style';
 
 interface HeaderSimpleProps {
   links: { link: string; label: string }[];
