+++ client/src/components/Register/Register.tsx
@@ -14,39 +14,7 @@ import {
   TextInput,
 } from '@mantine/core';
 import { useNavigate } from 'react-router-dom';
-
-const useStyles = createStyles((theme) => ({
-  wrapper: {
-    height: '100vh',
-    backgroundSize: 'cover',
-    backgroundImage: 'url(https://www.canossaschool.in/images/slide1.jpg)',
-    textAlign: 'left',
-  },
-
-  form: {
-    borderRight: `1px solid ${theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.colors.gray[3]}`,
-    height: '100vh',
-    maxWidth: 450,
-    paddingTop: 80,
-
-    [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-      maxWidth: '100%',
-    },
-  },
-
-  title: {
-    color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-    fontFamily: `Greycliff CF, ${theme.fontFamily}`,
-  },
-
-  logo: {
-    color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-    width: 120,
-    display: 'block',
-    marginLeft: 'auto',
-    marginRight: 'auto',
-  },
-}));
+import useStyles from './style';
 
 const AuthenticationForm = (props: PaperProps) => {
   const { classes } = useStyles();
